6.30_Phidata/requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
groq
langchain-groq
openai

#for playground
python-multipart
fastapi
uvicorn